Python 3 syntax changes
Update scitbx.matrix to python 3 syntax.

About half (25k) of the __init__.py consists of tests. To avoid having
to always loading tests when using the library move the tests out into
a separate file. Make sure starting either file directly will run the
tests.
